requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
openai>=1.50.0
pydantic-settings>=2.5.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
pytest>=8.0.0